Tried to do a bunch of integration tests but things went wrong
