+++ lib/godmin.rb
@@ -1,7 +1,9 @@
 require "bootstrap-sass"
 require "bootstrap_form"
 require "csv_builder"
+require "jquery-rails"
 require "momentjs-rails"
+require "sass-rails"
 require "selectize-rails"
 require "godmin/application_controller"
 require "godmin/authentication"
