+++ test/dummy/app/assets/stylesheets/application.css
@@ -8,6 +8,7 @@
  * You're free to add application-wide styles to this file and they'll appear at the top of the
  * compiled file, but it's generally better to create a new file per style scope.
  *
- *= require_self
+ *= require godmin
  *= require_tree .
+ *= require_self
  */
