+++ test/dummy/app/controllers/articles_controller.rb
@@ -1,0 +1,3 @@
+class ArticlesController < ApplicationController
+  include Godmin::Resources::ResourceController
+end
