+++ test/dummy/app/models/article.rb
@@ -1,0 +1,2 @@
+class Article < ActiveRecord::Base
+end
