+++ test/dummy/app/services/article_service.rb
@@ -1,0 +1,7 @@
+class ArticleService
+  include Godmin::Resources::ResourceService
+
+  attrs_for_index :id, :title, :published
+  attrs_for_show :id, :title, :body, :published
+  attrs_for_form :title, :body, :published
+end
