+++ test/dummy/app/views/articles/index.html.erb
@@ -1,0 +1,1 @@
+foobar
