+++ test/dummy/app/views/shared/_navigation.html.erb
@@ -1,0 +1,1 @@
+<%= navbar_item Article %>
