+++ test/dummy/config/routes.rb
@@ -1,3 +1,4 @@
 Rails.application.routes.draw do
+  resources :articles
   root to: "application#welcome"
 end
