+++ test/dummy/db/migrate/20150717121532_create_articles.rb
@@ -1,0 +1,11 @@
+class CreateArticles < ActiveRecord::Migration
+  def change
+    create_table :articles do |t|
+      t.string :title
+      t.text :body
+      t.boolean :published
+
+      t.timestamps null: false
+    end
+  end
+end
