+++ test/fakes/article.rb
@@ -1,0 +1,4 @@
+module Fakes
+  class Article
+  end
+end
